Move `Given` to a separate package.
